refactor(data): :recycle: move show error to seperate file
